+++ api/src/main/kotlin/org/kryptonmc/krypton/api/Server.kt
@@ -3,10 +3,12 @@ package org.kryptonmc.krypton.api
 import net.kyori.adventure.audience.ForwardingAudience
 import net.kyori.adventure.text.Component
 import org.kryptonmc.krypton.api.command.CommandManager
+import org.kryptonmc.krypton.api.command.Sender
 import org.kryptonmc.krypton.api.world.Gamemode
 import org.kryptonmc.krypton.api.entity.entities.Player
 import org.kryptonmc.krypton.api.event.EventBus
 import org.kryptonmc.krypton.api.plugin.PluginManager
+import org.kryptonmc.krypton.api.scheduling.Scheduler
 import org.kryptonmc.krypton.api.status.StatusInfo
 import org.kryptonmc.krypton.api.world.Difficulty
 import org.kryptonmc.krypton.api.world.WorldManager
@@ -53,6 +55,13 @@ interface Server : ForwardingAudience {
      * This is used to subscribe to and call events
      */
     val eventBus: EventBus
+
+    /**
+     * The scheduler for this server
+     *
+     * This can be used to run and schedule asynchronous tasks
+     */
+    val scheduler: Scheduler
 
     /**
      * The status information for this server
@@ -105,6 +114,11 @@ interface Server : ForwardingAudience {
      * @param name the player's name
      */
     fun player(name: String): Player?
+
+    /**
+     * The console's [Sender] object
+     */
+    val console: Sender
 
     /**
      * The server's [Scoreboard], or null if there isn't one