+++ api/src/main/kotlin/org/kryptonmc/krypton/api/event/events/play/MoveEvent.kt
@@ -1,16 +1,19 @@
 package org.kryptonmc.krypton.api.event.events.play
 
 import org.kryptonmc.krypton.api.entity.entities.Player
-import org.kryptonmc.krypton.api.event.CancellableEvent
+import org.kryptonmc.krypton.api.event.Event
 import org.kryptonmc.krypton.api.world.Location
 
 /**
  * Called when a player moves.
  *
+ * @param player the player who moved
+ * @param oldLocation the old location of the player
+ * @param newLocation the new location of the player
  * @author Callum Seabrook
  */
 class MoveEvent(
     val player: Player,
     val oldLocation: Location,
     val newLocation: Location
-) : CancellableEvent()
+) : Event