+++ server/src/main/kotlin/org/kryptonmc/krypton/console/ConsoleSender.kt
@@ -4,20 +4,24 @@ import net.kyori.adventure.audience.MessageType
 import net.kyori.adventure.identity.Identity
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer
-import org.apache.logging.log4j.LogManager
-import org.kryptonmc.krypton.api.command.Sender
+import org.kryptonmc.krypton.KryptonServer
 import org.kryptonmc.krypton.command.KryptonSender
+import org.kryptonmc.krypton.extension.logger
 
-class ConsoleSender : KryptonSender() {
+class ConsoleSender(server: KryptonServer) : KryptonSender(server) {
 
     override val name = "CONSOLE"
+
+    override fun identity() = Identity.nil()
 
     override fun sendMessage(source: Identity, message: Component, type: MessageType) {
         LOGGER.info(LegacyComponentSerializer.legacySection().serialize(message))
     }
 
+    override fun hasPermission(permission: String) = true // we are literally god, we never fail permission checks
+
     companion object {
 
-        private val LOGGER = LogManager.getLogger("CONSOLE")
+        private val LOGGER = logger("CONSOLE")
     }
 }