+++ server/src/main/kotlin/org/kryptonmc/krypton/entity/entities/KryptonPlayer.kt
@@ -3,6 +3,8 @@ package org.kryptonmc.krypton.entity.entities
 import net.kyori.adventure.audience.MessageType
 import net.kyori.adventure.identity.Identity
 import net.kyori.adventure.text.Component
+import net.kyori.adventure.title.Title
+import org.kryptonmc.krypton.KryptonServer
 import org.kryptonmc.krypton.api.effect.particle.ColorParticleData
 import org.kryptonmc.krypton.api.effect.particle.DirectionalParticleData
 import org.kryptonmc.krypton.api.effect.particle.NoteParticleData
@@ -12,19 +14,24 @@ import org.kryptonmc.krypton.api.entity.entities.Player
 import org.kryptonmc.krypton.api.space.Vector
 import org.kryptonmc.krypton.api.world.Gamemode
 import org.kryptonmc.krypton.api.world.Location
+import org.kryptonmc.krypton.api.world.World
 import org.kryptonmc.krypton.api.world.scoreboard.Scoreboard
 import org.kryptonmc.krypton.command.KryptonSender
 import org.kryptonmc.krypton.packet.out.play.PacketOutParticles
 import org.kryptonmc.krypton.packet.out.play.chat.PacketOutChat
+import org.kryptonmc.krypton.packet.out.play.chat.PacketOutPlayerListHeaderFooter
+import org.kryptonmc.krypton.packet.out.play.chat.PacketOutTitle
+import org.kryptonmc.krypton.packet.out.play.chat.TitleAction
 import org.kryptonmc.krypton.session.Session
 import java.net.InetSocketAddress
 import java.util.Locale
 import java.util.UUID
 
 class KryptonPlayer(
+    server: KryptonServer,
     val session: Session,
     override val address: InetSocketAddress = InetSocketAddress("127.0.0.1", 1)
-) : Player, KryptonSender() {
+) : Player, KryptonSender(server) {
 
     override lateinit var uuid: UUID
     override lateinit var name: String
@@ -46,9 +53,11 @@ class KryptonPlayer(
 
     override var scoreboard: Scoreboard? = null
 
-    override lateinit var locale: Locale
+    override var locale: Locale? = null
 
     var gamemode = Gamemode.SURVIVAL
+
+    override lateinit var world: World
 
     override fun spawnParticles(particleEffect: ParticleEffect, location: Location) {
         val packet = PacketOutParticles(particleEffect, location)
@@ -63,4 +72,28 @@ class KryptonPlayer(
     override fun sendMessage(source: Identity, message: Component, type: MessageType) {
         session.sendPacket(PacketOutChat(message, type, source.uuid()))
     }
+
+    override fun sendActionBar(message: Component) {
+        session.sendPacket(PacketOutTitle(TitleAction.SET_ACTION_BAR, actionBar = message))
+    }
+
+    override fun sendPlayerListHeaderAndFooter(header: Component, footer: Component) {
+        session.sendPacket(PacketOutPlayerListHeaderFooter(header, footer))
+    }
+
+    override fun showTitle(title: Title) {
+        session.sendPacket(PacketOutTitle(TitleAction.SET_TITLE, title))
+        session.sendPacket(PacketOutTitle(TitleAction.SET_SUBTITLE, title))
+        session.sendPacket(PacketOutTitle(TitleAction.SET_TIMES_AND_DISPLAY, title))
+    }
+
+    override fun clearTitle() {
+        session.sendPacket(PacketOutTitle(TitleAction.HIDE))
+    }
+
+    override fun resetTitle() {
+        session.sendPacket(PacketOutTitle(TitleAction.RESET))
+    }
+
+    override fun identity() = Identity.identity(uuid)
 }