+++ server/src/main/kotlin/org/kryptonmc/krypton/plugin/KryptonPluginManager.kt
@@ -10,8 +10,7 @@ import org.kryptonmc.krypton.extension.logger
 import java.io.File
 import java.io.FileNotFoundException
 import java.io.InputStreamReader
-import java.lang.Exception
-import java.net.URLClassLoader
+import java.lang.reflect.InvocationTargetException
 import java.nio.file.Path
 import java.util.jar.JarFile
 
@@ -19,10 +18,10 @@ class KryptonPluginManager(private val server: KryptonServer) : PluginManager {
 
     override val plugins = mutableSetOf<Plugin>()
 
-    init {
+    internal fun initialise() {
         File(ROOT_FOLDER, "plugins/").apply {
             mkdir()
-            walk().filter { !it.isDirectory }.forEach {
+            (listFiles() ?: return).filter { !it.isDirectory }.forEach {
                 val plugin = load(this, it) ?: return@forEach
 
                 val name = plugin.context.description.name
@@ -43,6 +42,9 @@ class KryptonPluginManager(private val server: KryptonServer) : PluginManager {
 
     override fun isInitialized(plugin: Plugin) = plugin in plugins && plugin.loadState == PluginLoadState.INITIALIZED
 
+    // if the cast here fails, something is seriously wrong
+    override fun addToClasspath(plugin: Plugin, path: Path) = (plugin.javaClass.classLoader as PluginClassLoader).addPath(path)
+
     private fun load(folder: File, file: File): Plugin? {
         val description = loadDescription(file)
         LOGGER.info("Loading ${description.name} version ${description.version}...")
@@ -57,51 +59,32 @@ class KryptonPluginManager(private val server: KryptonServer) : PluginManager {
             logger("[${description.name}]")
         )
 
-        val mainClassName = JarFile(file).use {
-            it.manifest.mainAttributes.getValue("Main-Class")
-                ?: throw IllegalStateException("The manifest does not contain a Main-Class attribute!")
-        }
-
-        val loader = URLClassLoader(arrayOf(file.toURI().toURL()))
+        val loader = PluginClassLoader(file.toURI().toURL())
 
         return try {
-            val jarClass = loader.loadClass(mainClassName)
+            val jarClass = loader.loadClass(description.main)
             val pluginClass = jarClass.asSubclass(Plugin::class.java)
             pluginClass.getDeclaredConstructor(PluginContext::class.java).newInstance(context)
-        } catch (exception: ClassNotFoundException) {
-            LOGGER.error("Could not find main class for plugin ${description.name}!")
-            LOGGER.info("Shutting down ${description.name} version ${description.version}...")
-            null
-        } catch (exception: ClassCastException) {
-            LOGGER.error("Main class of ${description.name} does not extend Plugin!")
-            LOGGER.info("Shutting down ${description.name} version ${description.version}")
-            null
-        } catch (exception: NoSuchMethodException) {
-            LOGGER.error("Main class of ${description.name} does not have a constructor that accepts a plugin context!")
-            LOGGER.info("Shutting down ${description.name} version ${description.version}")
-            null
         } catch (exception: Exception) {
-            LOGGER.error(
-                "An unexpected exception occured when attempting to load the main class of ${description.name}",
-                exception
-            )
-            LOGGER.info("Shutting down ${description.name} version ${description.version}")
+            LOGGER.error(when (exception) {
+                is ClassNotFoundException -> "Could not find main class for plugin ${description.name}!"
+                is ClassCastException -> "Main class of ${description.name} does not extend Plugin!"
+                is NoSuchMethodException -> "Main class of ${description.name} does not have a constructor that accepts a plugin context!"
+                is IllegalAccessException -> "Main class of ${description.name}'s primary constructor is not open!"
+                is InstantiationException -> "Main class of ${description.name} must not be abstract!"
+                is InvocationTargetException -> "Main class of ${description.name} threw an exception!"
+                else -> "An unexpected exception occurred when attempting to load the main class of ${description.name}"
+            }, exception)
+            LOGGER.info("Shutting down ${description.name} version ${description.version}...")
             null
         }
     }
 
-    private fun loadDescription(file: File): PluginDescriptionFile {
-        val jar = JarFile(file)
+    private fun loadDescription(file: File): PluginDescriptionFile = JarFile(file).use { jar ->
         val entry = jar.getJarEntry("plugin.conf")
             ?: throw FileNotFoundException("Plugin's JAR does not contain a plugin.conf!")
-
-        val stream = jar.getInputStream(entry)
-
-        try {
-            return HOCON.decodeFromConfig(ConfigFactory.parseReader(InputStreamReader(stream)))
-        } finally {
-            jar.close()
-            stream.close()
+        jar.getInputStream(entry).use {
+            HOCON.decodeFromConfig(ConfigFactory.parseReader(InputStreamReader(it)))
         }
     }